#1529 moved utility methods from factory class to util class added new SearchResponseUtil class
