+++ cobigen/cobigen-core-api/src/main/java/com/devonfw/cobigen/api/util/to/SearchResponseFactory.java
@@ -5,7 +5,6 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.glassfish.jersey.client.oauth2.OAuth2ClientSupport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -16,13 +15,6 @@ import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import jakarta.ws.rs.ProcessingException;
-import jakarta.ws.rs.client.Client;
-import jakarta.ws.rs.client.ClientBuilder;
-import jakarta.ws.rs.client.Invocation;
-import jakarta.ws.rs.client.WebTarget;
-import jakarta.ws.rs.core.Feature;
-import jakarta.ws.rs.core.MediaType;
-import jakarta.ws.rs.core.Response;
 
 /**
  * Handles the responses from various search REST API's
@@ -31,14 +23,14 @@ import jakarta.ws.rs.core.Response;
 public class SearchResponseFactory {
 
   /** Logger instance. */
-  private static final Logger LOG = LoggerFactory.getLogger(SearchResponseFactory.class);
+  static final Logger LOG = LoggerFactory.getLogger(SearchResponseFactory.class);
 
   /**
    * Gets a list of available search REST APIs (register a new search interface type here)
    *
    * @return list of available {@link SearchResponse}
    */
-  private static List<SearchResponse> getAvailableSearchInterfaces() {
+  static List<SearchResponse> getAvailableSearchInterfaces() {
 
     List<SearchResponse> availableSearchInterfaces = new ArrayList<>();
     availableSearchInterfaces.add(new MavenSearchResponse());
@@ -49,7 +41,7 @@ public class SearchResponseFactory {
   }
 
   /**
-   * Gets the download links by given repository type
+   * Gets the maven artifact download links by given base URL, groupId and optional authentication token
    *
    * @param baseURL String of the repository server URL
    * @param groupId the groupId to search for
@@ -86,78 +78,6 @@ public class SearchResponseFactory {
     }
 
     return downloadLinks;
-  }
-
-  /**
-   * Creates a @WebTarget with provided authentication token
-   *
-   * @param targetLink link to get response from
-   * @param token bearer token to use for authentication
-   * @return WebTarget to use as resource
-   */
-  public static WebTarget bearerAuthenticationWithOAuth2AtClientLevel(String targetLink, String token) {
-
-    Feature feature = OAuth2ClientSupport.feature(token);
-    Client client = ClientBuilder.newBuilder().register(feature).build();
-
-    WebTarget target = client.target(targetLink);
-    return target;
-  }
-
-  /**
-   * Gets a json response by given REST API target link using bearer authentication token
-   *
-   * @param targetLink link to get response from
-   * @param authToken bearer token to use for authentication
-   * @return String of json response
-   * @throws RESTSearchResponseException if the returned status code was not 200 OK
-   */
-  public static String getJsonResponseStringByTargetLink(String targetLink, String authToken)
-      throws RESTSearchResponseException {
-
-    WebTarget target = null;
-
-    if (authToken != null) {
-      target = bearerAuthenticationWithOAuth2AtClientLevel(targetLink, authToken);
-    } else {
-      Client client = ClientBuilder.newClient();
-      target = client.target(targetLink);
-    }
-
-    Response response = null;
-    Invocation.Builder request = target.request(MediaType.APPLICATION_JSON);
-    response = request.get();
-
-    int status = response.getStatus();
-    String jsonResponse = "";
-    if (status == 200) {
-      jsonResponse = response.readEntity(String.class);
-    } else {
-      throw new RESTSearchResponseException("The search REST API returned the unexpected status code: ",
-          String.valueOf(status));
-    }
-    return jsonResponse;
-  }
-
-  /**
-   * Creates a download link (concatenates maven repository link with groupId, artifact and version)
-   *
-   * @param mavenRepo link to the maven repository to use
-   * @param groupId for the download link
-   * @param artifactId for the download link
-   * @param version for the download link
-   * @param fileEnding file ending for the download link
-   * @return concatenated download link
-   * @throws MalformedURLException if the URL was not valid
-   */
-  public static URL createDownloadLink(String mavenRepo, String groupId, String artifactId, String version,
-      String fileEnding) throws MalformedURLException {
-
-    String parsedGroupId = groupId.replace(".", "/");
-    String downloadFile = artifactId + "-" + version + fileEnding;
-    String downloadLink = mavenRepo + "/" + parsedGroupId + "/" + artifactId + "/" + version + "/" + downloadFile;
-    URL url = new URL(downloadLink);
-    return url;
   }
 
 }
